Move function-level tests out to a separate file
